requirements.txt
celery[redis]~=5.3.1
Flask==3.0.0
waitress==2.1.2
python-dotenv~=0.21.0
py-healthcheck==1.10.1